Add HomeContentTab component for updating home content with images and videos
